+++ Auth-Server/src/main/java/com/dju/gdsc/AuthServerApplication.java
@@ -1,11 +1,10 @@
 package com.dju.gdsc;
 
-import com.dju.gdsc.global.properties.AppProperties;
+import com.dju.gdsc.domain.common.properties.AppProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-import org.springframework.security.oauth2.config.annotation.web.configuration.EnableAuthorizationServer;
 
 @EnableConfigurationProperties(AppProperties.class)
 @SpringBootApplication
