+++ Auth-Server/src/main/java/com/dju/gdsc/domain/common/config/LocalCacheConfig.java
@@ -1,0 +1,4 @@
+package com.dju.gdsc.domain.common.config;
+
+public class LocalCacheConfig {
+}
