+++ Auth-Server/src/main/java/com/dju/gdsc/domain/common/config/corsConfig/CorsConfigNotReal.java
@@ -1,4 +1,4 @@
-package com.dju.gdsc.global.corsConfig;
+package com.dju.gdsc.domain.common.config.corsConfig;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -6,19 +6,16 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 import org.springframework.web.filter.CorsFilter;
+
+@Profile("!real")
 @Configuration
-@Profile("real")
-public class CorsConfigReal implements CorsConfig {
-    /**
-     * @return
-     */
+public class CorsConfigNotReal implements CorsConfig {
     @Bean
     public CorsFilter corsFilter() {
         UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
         CorsConfiguration config = new CorsConfiguration();
         config.setAllowCredentials(true); //  내 서버가 응답할때 json을 자바스크립트가 처리할수있게 할건지
-
-        config.addAllowedOriginPattern("https://gdsc-dju-blog.web.app");
+        config.addAllowedOriginPattern("*");
         config.addAllowedHeader("*"); // 모든 헤더 응답허용
         config.addAllowedMethod("*"); // 모든 post get put delete 요청 허용
         source.registerCorsConfiguration("/**", config);
