+++ Auth-Server/src/main/java/com/dju/gdsc/domain/common/dto/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.dju.gdsc.global.dto;
+package com.dju.gdsc.domain.common.dto;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
