+++ Auth-Server/src/main/java/com/dju/gdsc/domain/oauth/controller/AuthController.java
@@ -11,9 +11,9 @@ import com.dju.gdsc.domain.oauth.repository.UserRefreshTokenRepository;
 import com.dju.gdsc.domain.oauth.token.AuthToken;
 import com.dju.gdsc.domain.oauth.token.AuthTokenProvider;
 import com.dju.gdsc.domain.oauth.utils.CookieUtil;
-import com.dju.gdsc.global.dto.ApiResponse;
-import com.dju.gdsc.global.dto.ResponseDto;
-import com.dju.gdsc.global.properties.AppProperties;
+import com.dju.gdsc.domain.common.dto.ApiResponse;
+import com.dju.gdsc.domain.common.dto.ResponseDto;
+import com.dju.gdsc.domain.common.properties.AppProperties;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Profile;
