+++ Auth-Server/src/main/java/com/dju/gdsc/domain/oauth/controller/RefreshController.java
@@ -7,8 +7,8 @@ import com.dju.gdsc.domain.oauth.repository.UserRefreshTokenRepository;
 import com.dju.gdsc.domain.oauth.token.AuthToken;
 import com.dju.gdsc.domain.oauth.token.AuthTokenProvider;
 import com.dju.gdsc.domain.oauth.utils.HeaderUtil;
-import com.dju.gdsc.global.dto.ApiResponse;
-import com.dju.gdsc.global.properties.AppProperties;
+import com.dju.gdsc.domain.common.dto.ApiResponse;
+import com.dju.gdsc.domain.common.properties.AppProperties;
 import io.jsonwebtoken.Claims;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
