+++ Auth-Server/src/main/java/com/dju/gdsc/domain/oauth/handler/OAuth2AuthenticationSuccessHandler.java
@@ -11,8 +11,7 @@ import com.dju.gdsc.domain.oauth.repository.UserRefreshTokenRepository;
 import com.dju.gdsc.domain.oauth.token.AuthToken;
 import com.dju.gdsc.domain.oauth.token.AuthTokenProvider;
 import com.dju.gdsc.domain.oauth.utils.CookieUtil;
-import com.dju.gdsc.global.properties.AppProperties;
-import com.nimbusds.oauth2.sdk.http.HTTPResponse;
+import com.dju.gdsc.domain.common.properties.AppProperties;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.Authentication;
